+++ agent/src/index.test.ts
@@ -13,7 +13,12 @@ import { URI } from 'vscode-uri'
 import { TestClient, asTranscriptMessage } from './TestClient'
 import { decodeURIs } from './decodeURIs'
 import { isNode16 } from './isNode16'
-import type { CustomChatCommandResult, CustomEditCommandResult, EditTask } from './protocol-alias'
+import type {
+    CustomChatCommandResult,
+    CustomEditCommandResult,
+    EditTask,
+    Requests,
+} from './protocol-alias'
 
 const explainPollyError = `
 
@@ -1404,6 +1409,50 @@ describe('Agent', () => {
             },
             30_000
         )
+
+        it('remoteRepo/list', async () => {
+            // List a repo without a query
+            let repos: Requests['remoteRepo/list'][1]
+            do {
+                repos = await enterpriseClient.request('remoteRepo/list', {
+                    query: undefined,
+                    first: 10,
+                })
+            } while (repos.state.state === 'fetching')
+            expect(repos.repos).toHaveLength(10)
+
+            // Make a paginated query.
+            const secondLastRepo = repos.repos.at(-2)
+            const moreRepos = await enterpriseClient.request('remoteRepo/list', {
+                query: undefined,
+                first: 2,
+                afterId: secondLastRepo?.id,
+            })
+            expect(moreRepos.repos[0].id).toBe(repos.repos.at(-1)?.id)
+
+            // Make a query.
+            const filteredRepos = await enterpriseClient.request('remoteRepo/list', {
+                query: 'sourceco',
+                first: 1000,
+            })
+            expect(
+                filteredRepos.repos.find(repo => repo.name === 'github.com/sourcegraph/cody')
+            ).toBeDefined()
+        })
+
+        it('remoteRepo/has', async () => {
+            // Query a repo that does exist.
+            const codyRepoExists = await enterpriseClient.request('remoteRepo/has', {
+                repoName: 'github.com/sourcegraph/cody',
+            })
+            expect(codyRepoExists.result).toBe(true)
+
+            // Query a repo that does not exist.
+            const codyForDos = await enterpriseClient.request('remoteRepo/has', {
+                repoName: 'github.com/sourcegraph/cody-edlin',
+            })
+            expect(codyForDos.result).toBe(false)
+        })
 
         afterAll(async () => {
             const { requests } = await enterpriseClient.request('testing/networkRequests', null)