+++ vscode/src/chat/chat-view/SimpleChatPanelProvider.ts
@@ -8,7 +8,6 @@ import {
     type ContextItem,
     type DefaultChatCommands,
     type EventSource,
-    FeatureFlag,
     type FeatureFlagProvider,
     type Guardrails,
     MAX_BYTES_PER_FILE,
@@ -18,7 +17,6 @@ import {
     type SerializedChatInteraction,
     type SerializedChatTranscript,
     Typewriter,
-    featureFlagProvider,
     hydrateAfterPostMessage,
     isDefined,
     isError,
@@ -402,10 +400,6 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
         command?: DefaultChatCommands
     ): Promise<void> {
         return tracer.startActiveSpan('chat.submit', async (span): Promise<void> => {
-            const useFusedContextPromise = featureFlagProvider.evaluateFeatureFlag(
-                FeatureFlag.CodyChatFusedContext
-            )
-
             const authStatus = this.authProvider.getAuthStatus()
             const sharedProperties = {
                 requestID,
@@ -467,10 +461,6 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
                                       localEmbeddings: this.localEmbeddings,
                                       symf: this.config.experimentalSymfContext ? this.symf : null,
                                       remoteSearch: this.remoteSearch,
-                                  },
-                                  featureFlags: {
-                                      fusedContext:
-                                          this.config.internalUnstable || (await useFusedContextPromise),
                                   },
                                   hints: { maxChars },
                                   contextRanking: this.contextRanking,